Preface

Preface is an opinionated library designed to facilitate the handling of recurring functional programming idioms in OCaml. Many of the design decisions were made in an attempt to calibrate, as best as possible, to the OCaml language. Trying to get the most out of the module language. The name "Preface" is a nod to "Prelude".

When learning functional programming, one is often confronted with constructs derived (or not) from category theory. Languages such as Haskell offer very complete libraries to use them, and thus, facilitate their learning. In OCaml, it often happens that these abstractions are buried in the heart of certain libraries/projects. This is why one of the objectives of Preface is to propose tools to achieve these abstractions, at least as a pedagogical tool.

Preface on Github

Bugtracker

The library is released under the MIT License

Full Documentation

You can browse the full library documentation by clicking on the link below.

Access to the full library API.